refactor(core): split module map data out
